Phase-1-Console-App/pyproject.toml
[project]
name = "evolution-of-todo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "plyer>=2.1.0",
    "pyaudio>=0.2.14",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "speechrecognition>=3.14.4",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
